src/requirements.txt
python-binance
pandas
numpy
pandas_ta
ta
requests
pydantic
sqlalchemy
mysql-connector-python
icecream